work in progress
iOS tools and IntelliJ installation working. Corretto in progress, checking in now because I want to keep a record of some code I’m about to delete so I don’t have to recreated it from scratch if I need to come back to it. (Specifically, code around disabling a web Java plug-in so as to try to avoid the need to brute-force it by setting JAVA_HOME.)